Rename view-all-books page
